requirements.txt
Django==4.2.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.3.1
Pillow==10.1.0
tensorflow-cpu==2.15.0
numpy==1.26.0
opencv-python==4.8.1.78
requests==2.31.0
python-decouple==3.8
gunicorn==21.2.0
whitenoise==6.6.0
